We have done a Windows Form application with 2 classroms.
In the first one we put 4 students, in the second one is the same.
If you click on the first classroom it shows the members of the first classroom and work equals with the second classroom.
When you click on a student it shows on the textBox, and finally if you click on Clear list it deletes the classroom.